# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shopmee API Documentation",
        "description": "API endpoints for Shopmee e-commerce application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Shopmee API Server"
        }
    ],
    "paths": {
...
        "/orders/{id}/status": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Kiểm tra trạng thái đơn hàng",
                "description": "Kiểm tra trạng thái thanh toán và giao hàng hiện tại của đơn hàng.",
                "operationId": "checkOrderStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của đơn hàng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "payment_status": {
                                            "type": "string",
                                            "example": "paid"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "processing"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy đơn hàng"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/seller/dashboard/revenue": {
            "get": {
                "tags": [
                    "Seller Dashboard"
                ],
                "summary": "Xem tổng doanh thu và thống kê của Shop",
                "description": "API thống kê doanh thu theo mốc thời gian dành cho Seller.",
                "operationId": "getSellerRevenue",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Ngày bắt đầu (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Ngày kết thúc (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "description": "Ngày bắt đầu thay thế (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "description": "Ngày kết thúc thay thế (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "total_revenue": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1548000
                                                },
                                                "orders_count": {
                                                    "type": "integer",
                                                    "example": 12
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Tài khoản không phải Seller hoặc chưa đăng ký Shop"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/seller/orders/{id}": {
            "patch": {
                "tags": [
                    "Seller Orders"
                ],
                "summary": "Cập nhật trạng thái đơn hàng (dành cho Seller)",
                "description": "Seller cập nhật trạng thái đơn hàng của shop (confirmed, shipping, delivered, cancelled).",
                "operationId": "updateOrderStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của đơn hàng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "confirmed",
                                        "enum": [
                                            "confirmed",
                                            "shipping",
                                            "delivered",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cập nhật trạng thái đơn hàng thành công"
                                        },
                                        "data": {
                                            "properties": {
                                                "order_id": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "confirmed"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Yêu cầu không hợp lệ"
                    },
                    "403": {
                        "description": "Không có quyền chỉnh sửa đơn hàng này"
                    },
                    "404": {
                        "description": "Không tìm thấy đơn hàng"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}